+++ src/components/ComboBoxButton/ComboBoxButton.component.tsx
@@ -1,11 +1,14 @@
 import { isNull } from "@ubloimmo/front-util";
-import { MouseEventHandler, useCallback, useMemo } from "react";
+import { MouseEventHandler, useCallback, useMemo, useState } from "react";
 import styled from "styled-components";
 
 import {
   ComboBoxButtonStyles,
+  ComboBoxButtonWrapperStyles,
+  ComboBoxContextMenuTriggerStyles,
   ComboBoxIconContainerStyle,
 } from "./ComboBoxButton.styles";
+import { ContextMenu, ContextMenuItemData } from "../ContextMenu";
 import { Icon, type IconProps } from "../Icon";
 import { Text } from "../Text";
 
@@ -35,12 +38,18 @@ const defaultComboBoxButtonProps: ComboBoxButtonDefaultProps = {
   fill: false,
   showIcon: true,
   description: null,
+  deletable: false,
+  editable: false,
+  onEdit: null,
+  onDelete: null,
+  editLabel: null,
+  deleteLabel: null,
 };
 
 /**
  * A single clickable option in a ComboBox
  *
- * @version 0.0.5
+ * @version 0.0.6
  *
  * @param {ComboBoxButtonProps & TestIdProps} props - ComboBoxButton component props
  * @returns {JSX.Element}
@@ -50,9 +59,27 @@ const ComboBoxButton = (
 ): JSX.Element => {
   const { warn } = useLogger("ComboBoxButton");
   const mergedProps = useMergedProps(defaultComboBoxButtonProps, props);
-  const { label, multi, active, showIcon, disabled, description } = mergedProps;
-  const styleProps = useStyleProps(mergedProps);
   const testId = useTestId("combo-box-button", props);
+  const styleProps = useStyleProps<ComboBoxButtonDefaultProps & TestIdProps>({
+    ...mergedProps,
+    testId,
+  });
+  const {
+    label,
+    multi,
+    active,
+    showIcon,
+    disabled,
+    description,
+    editable,
+    deletable,
+    onDelete,
+    onEdit,
+    deleteLabel,
+    editLabel,
+  } = mergedProps;
+
+  const [menuOpen, setMenuOpen] = useState(false);
 
   if (!props.label || isEmptyString(props.label)) {
     warn(
@@ -112,45 +139,87 @@ const ComboBoxButton = (
     return tl.action[active ? "unselect" : "select"](label);
   }, [active, label, tl.action]);
 
-  return (
-    <ComboBoxButtonContainer
-      data-testid={testId}
-      onClick={onSelect}
-      disabled={mergedProps.disabled}
-      type="button"
-      title={ariaLabel}
-      aria-label={ariaLabel}
-      aria-disabled={mergedProps.disabled}
-      {...styleProps}
-    >
-      {showIcon && (
-        <ComboBoxIconContainer $active={active ?? false}>
-          <Icon {...iconProps} />
-          <Icon {...iconProps} />
-        </ComboBoxIconContainer>
-      )}
+  const shouldDisplayContextMenu = useMemo(
+    () => (editable || deletable) && !disabled,
+    [editable, deletable, disabled]
+  );
 
-      <FlexColumnLayout
-        gap="s-1"
-        align="start"
-        justify="start"
-        testId="combo-box-button-content"
-        overrideTestId
+  const contextMenuItems = useMemo<ContextMenuItemData[]>(
+    () => [
+      {
+        label: editLabel ?? tl.action.edit(),
+        disabled: disabled || !editable,
+        onClick: onEdit,
+      },
+      {
+        label: deleteLabel ?? tl.action.delete(),
+        disabled: disabled || !deletable,
+        onClick: onDelete,
+      },
+    ],
+    [
+      deletable,
+      deleteLabel,
+      disabled,
+      editLabel,
+      editable,
+      onDelete,
+      onEdit,
+      tl.action,
+    ]
+  );
+
+  return (
+    <ComboBoxButtonWrapper {...styleProps} role="listitem">
+      <ComboBoxButtonContainer
+        data-testid={testId}
+        onClick={onSelect}
+        disabled={mergedProps.disabled}
+        type="button"
+        title={ariaLabel}
+        aria-label={ariaLabel}
+        aria-disabled={mergedProps.disabled}
       >
-        <Text {...textProps} testId="combo-box-button-label" overrideTestId>
-          {label}
-        </Text>
-        {description && (
-          <Text
-            color="gray-600"
-            testId="combo-box-button-description"
-            overrideTestId
-          >
-            {description}
-          </Text>
+        {showIcon && (
+          <ComboBoxIconContainer $active={active ?? false}>
+            <Icon {...iconProps} />
+            <Icon {...iconProps} />
+          </ComboBoxIconContainer>
         )}
-      </FlexColumnLayout>
-    </ComboBoxButtonContainer>
+
+        <FlexColumnLayout
+          gap="s-1"
+          align="start"
+          justify="start"
+          testId="combo-box-button-content"
+          overrideTestId
+        >
+          <Text {...textProps} testId="combo-box-button-label" overrideTestId>
+            {label}
+          </Text>
+          {description && (
+            <Text
+              color="gray-600"
+              testId="combo-box-button-description"
+              overrideTestId
+            >
+              {description}
+            </Text>
+          )}
+        </FlexColumnLayout>
+      </ComboBoxButtonContainer>
+      {shouldDisplayContextMenu && (
+        <ContextMenu
+          open={menuOpen}
+          onOpenChange={setMenuOpen}
+          items={contextMenuItems}
+        >
+          <ComboBoxContextMenuTrigger type="button">
+            <Icon name="ThreeDotsVertical" color={iconProps.color} />
+          </ComboBoxContextMenuTrigger>
+        </ContextMenu>
+      )}
+    </ComboBoxButtonWrapper>
   );
 };
 
@@ -158,12 +227,20 @@ ComboBoxButton.defaultProps = defaultComboBoxButtonProps;
 
 export { ComboBoxButton };
 
-const ComboBoxButtonContainer = styled.button<
-  StyleProps<ComboBoxButtonDefaultProps>
->`
+const ComboBoxButtonContainer = styled.button`
   ${ComboBoxButtonStyles}
 `;
 
 const ComboBoxIconContainer = styled.div<ComboButtonIconContainerStyleProps>`
   ${ComboBoxIconContainerStyle}
+`;
+
+const ComboBoxButtonWrapper = styled.article<
+  StyleProps<ComboBoxButtonDefaultProps & TestIdProps>
+>`
+  ${ComboBoxButtonWrapperStyles}
+`;
+
+const ComboBoxContextMenuTrigger = styled.button`
+  ${ComboBoxContextMenuTriggerStyles}
 `;