+++ src/components/ComboBoxButton/ComboBoxButton.styles.ts
@@ -7,63 +7,102 @@ import type {
   ComboBoxButtonDefaultProps,
   ComboButtonIconContainerStyleProps,
 } from "./ComboBoxButton.types";
-import type { StyleProps } from "@types";
+import type { StyleProps, TestIdProps } from "@types";
 
-export const ComboBoxButtonStyles = ({
+export const ComboBoxButtonWrapperStyles = ({
   $active,
   $fill,
   $description,
-}: StyleProps<ComboBoxButtonDefaultProps>): RuleSet => {
-  const height = $description
+  $testId,
+}: StyleProps<ComboBoxButtonDefaultProps & TestIdProps>): RuleSet => css`
+  --combobox-button-background: ${$active
+    ? cssVarUsage("primary-light")
+    : "white"};
+  --combobox-button-border-color: var(--primary-medium);
+  --combobox-button-min-height: var(--s-8);
+  --combobox-button-height: ${$description
     ? "fit-content"
-    : cssVarUsage("combobox-button-height");
+    : cssVarUsage("combobox-button-min-height")};
+
+  display: inline-grid;
+  grid-template-columns: max-content auto;
+  grid-template-rows: 1fr;
+  background: none;
+  border-radius: var(--s-1);
+
+  ${$fill &&
+  css`
+    flex: 1;
+
+    button:first-child {
+      flex: 1;
+      width: 100%;
+      height: 100%;
+      max-height: 100%;
+    }
+  `}
+
+  &:not(:has(button[data-testid="${$testId}"]:disabled:first-child)) {
+    box-shadow: var(--shadow-button);
+  }
+
+  &:has(button[data-testid="${$testId}"]:disabled:first-child) {
+    --combobox-button-background: var(--gray-50);
+    --combobox-button-border-color: var(--gray-300);
+  }
+
+  &:not(button[data-testid="${$testId}"]:disabled:first-child):hover {
+    --combobox-button-border-color: var(--primary-base);
+  }
+
+  &:has(div[type="button"]) {
+    & > button[data-testid="${$testId}"]:first-child {
+      border-top-right-radius: 0;
+      border-bottom-right-radius: 0;
+    }
+
+    & > div[type="button"] {
+      min-height: var(--combobox-button-min-height);
+      height: 100%;
+    }
+  }
+
+  &:not(:has(div[type="button"])) > button[data-testid="${$testId}"] {
+    grid-column-end: -1;
+  }
+
+  @media only screen and (max-width: ${breakpointsPx.XS}) {
+    --combobox-button-height: var(--s-10);
+  }
+`;
+
+export const ComboBoxButtonStyles = (): RuleSet => {
   return css`
-    --combobox-button-height: var(--s-8);
-    min-height: var(--combobox-button-height);
-    height: ${height};
-    max-height: ${height};
+    min-height: var(--combobox-button-min-height);
+    height: var(--combobox-button-height);
+    max-height: var(--combobox-button-height);
     width: max-content;
     display: flex;
     align-items: center;
     justify-content: flex-start;
     gap: var(--s-2);
     cursor: pointer;
-    background: ${$active ? cssVarUsage("primary-light") : "white"};
+    background: var(--combobox-button-background);
     padding: var(--s-2) var(--s-4);
     border-radius: var(--s-1);
-    border: 1px solid var(--primary-medium);
+    border: 1px solid var(--combobox-button-border-color);
     transition: 300ms background ease-out 0s, 300ms border-color ease-out 0s;
     overflow: hidden;
-
-    ${$fill &&
-    css`
-      flex: 1;
-    `}
+    grid-column: 1 / span 1;
+    grid-row: 1 / span 1;
 
     span[data-testid="combo-box-button-label"] {
       overflow: hidden;
     }
 
-    &:not(:disabled) {
-      box-shadow: var(--shadow-button);
-    }
-
-    &:disabled {
-      background: var(--gray-50);
-      border-color: var(--gray-300);
-    }
-
-    &:hover:not(:disabled) {
-      border-color: var(--primary-base);
-    }
-
     div[data-testid="combo-box-button-content"] {
       flex: 1;
       overflow: hidden;
-    }
-
-    @media only screen and (max-width: ${breakpointsPx.XS}) {
-      --combobox-button-height: var(--s-10);
     }
   `;
 };
@@ -100,3 +139,22 @@ export const ComboBoxIconContainerStyle = ({
     }
   `;
 };
+
+export const ComboBoxContextMenuTriggerStyles = (): RuleSet => css`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+  width: var(--s-8);
+  min-height: var(--combobox-button-min-height);
+  border: 1px solid var(--combobox-button-border-color);
+  border-left: none;
+  background: var(--combobox-button-background);
+  border-top-right-radius: var(--s-1);
+  border-bottom-right-radius: var(--s-1);
+  transition: 300ms background ease-out 0s, 300ms border-color ease-out 0s;
+
+  grid-column: 2 / span 1;
+  grid-row: 1 / span 1;
+  cursor: pointer;
+`;